+++ src/main/java/hanium/smath/Member/controller/MemberController.java
@@ -1,18 +1,14 @@
 package hanium.smath.Member.controller;
 
-import hanium.smath.Member.dto.KakaoLoginRequest; // KakaoLoginRequest 추가
-import hanium.smath.Member.dto.GoogleLoginRequest;
-import hanium.smath.Member.dto.LoginResponse;
+import hanium.smath.Member.dto.*;
 import hanium.smath.Member.repository.EmailVerificationRepository;
-import hanium.smath.Member.dto.KakaoProfile;
-import hanium.smath.Member.dto.LoginRequest;
 import hanium.smath.Member.entity.EmailVerification;
 import hanium.smath.Member.entity.Member;
 import hanium.smath.Member.security.JwtUtil;
 import hanium.smath.Member.service.EmailService;
 import hanium.smath.Member.service.GoogleLoginService;
 import hanium.smath.Member.service.KakaoService;
-import hanium.smath.Member.service.LoginService;
+import hanium.smath.Member.service.MemberService;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -22,14 +18,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.service.annotation.GetExchange;
-
 
 @RestController
 @RequestMapping("/api/members")
-public class LoginController {
+public class MemberController {
 
-    private LoginService loginService;
+    private final MemberService memberService;
     private final GoogleLoginService googleLoginService;
     private final JwtUtil jwtUtil;
     private final EmailService emailService;
@@ -38,8 +34,8 @@ public class LoginController {
     private final PasswordEncoder passwordEncoder;
 
     @Autowired
-    public LoginController(LoginService loginService, GoogleLoginService googleLoginService, JwtUtil jwtUtil, EmailService emailService, KakaoService kakaoService, EmailVerificationRepository emailVerificationRepository, PasswordEncoder passwordEncoder) {
-        this.loginService = loginService; // controller가 생성될 때 service 주입하기
+    public MemberController(MemberService memberService, GoogleLoginService googleLoginService, JwtUtil jwtUtil, EmailService emailService, KakaoService kakaoService, EmailVerificationRepository emailVerificationRepository, PasswordEncoder passwordEncoder) {
+        this.memberService = memberService; // controller가 생성될 때 service 주입하기
         System.out.println("MemberController instantiated with MemberService");
         this.googleLoginService = googleLoginService;
         System.out.println("MemberController instantiated with googleLoginService");
@@ -53,6 +49,118 @@ public class LoginController {
         System.out.println("MemberController instantiated with emailVerificationRepository");
         this.passwordEncoder = passwordEncoder;
     }
+
+    //회원가입
+    //아이디 중복 확인
+    @PostMapping("/check-loginId")
+    public ResponseEntity<String> checkLoginId(@RequestParam String loginId) {
+        if (memberService.checkLoginIdExists(loginId)) {
+            return ResponseEntity.badRequest().body("Login ID already exists");
+        } else {
+            return ResponseEntity.ok("Login ID is available");
+        }
+    }
+
+    //이메일 중복 확인 후 인증 메일 전송
+    @PostMapping("/check-email")
+    public ResponseEntity<String> checkEmail(@RequestParam String email) {
+        if (memberService.checkEmailExists(email)) {
+            return ResponseEntity.badRequest().body("Email already exists");
+        } else {
+            // 새로운 인증 코드 생성
+            int newCode = emailService.generateVerificationCode();
+
+            // 이메일로 기존의 인증 레코드를 가져옴
+            Optional<EmailVerification> optionalVerification = emailVerificationRepository.findTopByEmailOrderByCreateTimeDesc(email);
+
+            if (optionalVerification.isPresent()) {
+                // 기존 레코드가 있으면 인증번호만 업데이트
+                EmailVerification existingVerification = optionalVerification.get();
+                existingVerification.setVerificationCode(newCode);
+                existingVerification.setCreateTime(LocalDateTime.now()); // 생성 시간 갱신
+                emailVerificationRepository.save(existingVerification);
+                System.out.println("EmailController: Updated verification code for email: " + email);
+
+                // 인증 코드 이메일로 전송
+                emailService.sendVerificationEmail(email, newCode);
+
+                return ResponseEntity.ok("Verification code updated and sent to: " + email);
+            } else {
+                memberService.sendVerificationCodeToEmail(email);
+                return ResponseEntity.ok("Email is available. Verification code sent to email.");
+            }
+        }
+    }
+
+    //이메일 검증
+    @PatchMapping("/verify-email")
+    public ResponseEntity<String> verifyEmail(@RequestParam String email, @RequestParam String code) {
+        int codeInt = Integer.parseInt(code);
+        boolean isVerified = memberService.verifyEmailCode(email, codeInt);
+
+        if (isVerified) {
+            return ResponseEntity.ok("Email verified successfully");
+        } else {
+            return ResponseEntity.badRequest().body("Invalid verification code");
+        }
+    }
+
+    @PostMapping("/signup")
+    public ResponseEntity<SignupResponse> registerMember(@RequestBody SignupRequest signupRequest) {
+        // 이메일 인증 정보 가져오기
+        EmailVerification optionalEmailVerification = emailVerificationRepository.findEmailVerificationByEmail(signupRequest.getEmail());
+
+        if (optionalEmailVerification == null || !optionalEmailVerification.isVerifiedEmail()) {
+            return ResponseEntity.badRequest().body(null);
+        }
+
+        if (memberService.checkLoginIdExists(signupRequest.getLoginId())) {
+            return ResponseEntity.badRequest().body(null);
+        }
+
+        // 비밀번호 암호화
+        String encodedPassword = passwordEncoder.encode(signupRequest.getLoginPwd());
+        signupRequest.setLoginPwd(encodedPassword);
+
+        // 회원 등록
+        memberService.registerMember(signupRequest);
+
+        // 회원 정보 가져오기 - 회원가입 후 자동 로그인
+        Member member = memberService.getMemberById(signupRequest.getLoginId());
+        if (member == null) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
+        }
+
+        // JWT 토큰 생성
+        String token = jwtUtil.generateToken(member.getLoginId());
+
+        // 응답 반환 - 토큰만 반환
+        SignupResponse response = new SignupResponse(token);
+        return ResponseEntity.ok(response);
+    }
+
+    //회원 삭제
+    @DeleteMapping("/delete")
+    public ResponseEntity<String> deleteMember() {
+        memberService.deleteCurrentUser();
+        return ResponseEntity.ok("Your account has been deleted.");
+    }
+
+    //닉네임 변경
+    @PatchMapping("/change_nickname")
+    public ResponseEntity<String> changeNickname(@RequestBody NicknameChangeRequest request) {
+        memberService.changeNickname(request.getNewNickname());
+        return ResponseEntity.ok("Nickname has been changed successfully.");
+    }
+
+    //아이콘 변경
+    @PatchMapping("/change_icon")
+    public ResponseEntity<String> changeIcon(@RequestBody IconChangeRequest request) {
+        memberService.changeIcon(request.getNewIcon());
+        return ResponseEntity.ok("Icon has been changed successfully.");
+    }
+
+
 
     // 로그인
     @PostMapping("/login")
@@ -71,7 +179,7 @@ public class LoginController {
 
         try {
             System.out.println("Login request received for ID: " + loginRequest.getLogin_id());
-            Member member = loginService.getMemberById(loginRequest.getLogin_id());
+            Member member = memberService.getMemberById(loginRequest.getLogin_id());
 
             if (member != null) {
                 System.out.println("Member found : " + loginRequest.getLogin_id());
@@ -135,7 +243,7 @@ public class LoginController {
             KakaoProfile kakaoProfile = kakaoService.getKakaoProfile(request.getAccessToken());
 
             // 사용자 정보로 로그인 처리
-            Member member = loginService.processKakaoLogin(kakaoProfile);
+            Member member = memberService.processKakaoLogin(kakaoProfile);
 
             if (member != null) {
                 // JWT 토큰 생성
@@ -161,7 +269,7 @@ public class LoginController {
         System.out.println("Received request to find login ID");
         try {
             LocalDate birthDateLocal = LocalDate.parse(birthdate);
-            String loginId = loginService.findLoginId(email, birthDateLocal);
+            String loginId = memberService.findLoginId(email, birthDateLocal);
             System.out.println("Found ID: " + loginId);
             return ResponseEntity.ok(loginId);  // 로그인 ID를 직접 반환
         } catch (RuntimeException e) {
@@ -203,7 +311,6 @@ public class LoginController {
         }
     }
 
-
     // 이메일로 보낸 인증 코드를 검증하는 API
     @GetMapping("/find/email/verify")
     public ResponseEntity<String> verifyCodeFindId(@RequestParam String email, @RequestParam String code) {
@@ -223,7 +330,7 @@ public class LoginController {
     public ResponseEntity<String> initiatePasswordReset(@RequestParam String login_id) {
         System.out.println("LoginController: Initiating password reset for loginId: " + login_id);
         try {
-            Member member = loginService.getMemberById(login_id);
+            Member member = memberService.getMemberById(login_id);
             if (member != null) {
                 // 회원이 존재하면 이메일을 반환
                 String email = member.getEmail();
@@ -251,7 +358,7 @@ public class LoginController {
         }
 
         // 비밀번호 변경
-        boolean passwordChanged = loginService.changeUserPassword(email, new_password);
+        boolean passwordChanged = memberService.changeUserPassword(email, new_password);
         if (passwordChanged) {
             // 인증 코드 무효화 (실제로는 verifiedEmail을 true로 설정)
             emailService.invalidateVerificationCode(email, code);
