+++ src/main/java/hanium/smath/Member/entity/Member.java
@@ -6,7 +6,6 @@ import org.hibernate.annotations.CreationTimestamp;
 
 import java.sql.Timestamp;
 import java.time.LocalDate;
-
 @Getter
 @Setter
 @NoArgsConstructor
